src-python/requirements.txt
# NeoView Python Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
orjson>=3.9.0
aiofiles>=23.2.0
natsort>=8.4.0
pillow>=10.2.0
PyTurboJPEG>=1.7.0
rarfile>=4.1
py7zr>=0.20.0
av>=11.0.0
send2trash>=1.8.0
cachetools>=5.3.0
ebooklib>=0.18

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0
httpx>=0.26.0
